builder_requirements.txt
absl-py==0.1.12
alabaster==0.7.10
appdirs==1.4.3
astor==0.6.2
astroid==1.6.2
attrs==17.4.0
Babel==2.5.3
bleach==1.5.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
ConfigArgParse==0.13.0
coverage==4.5.1
docutils==0.14
filelock==3.0.4
flake8==3.5.0
future==0.16.0
gast==0.2.0
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
imagesize==1.0.0
isort==4.3.4
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
nervananeon==2.6.0
numpy==1.14.2
packaging==17.1
pandoc==1.0.2
pep8==1.7.1
Pillow==5.0.0
pluggy==0.6.0
ply==3.11
posix-ipc==1.0.4
protobuf==3.5.2.post1
py==1.5.2
py-cpuinfo==3.3.0
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.3
pypandoc==1.4
pyparsing==2.2.0
pytest==3.4.2
pytest-cov==2.5.1
pytest-mock==1.7.1
pytz==2018.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.2
sphinxcontrib-websupport==1.0.1
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
tqdm==4.19.8
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
